merges bootstrap_monitors with monitors files
