+++ src/ui/src/pages/app/monitors/Index/components/BootstrapMonitorsTable/index.tsx
