+++ src/ui/src/pages/app/monitors/Index/components/BootstrapMonitorsTable/table.css
